docs(preprocess-twmerge): document code chore(preprocess-twmerge): clean up code and break ast library into separate files
